Updated blackhole, bowling ball, TNT, spawning, tags, player controller

+++ Assets/Scripts/PlayerController.cs
@@ -35,8 +35,8 @@ public class PlayerController : MonoBehaviour
     {
         if (!gameLoopHandler.gameOver)
         {
-            int hor = Convert.ToInt16(Input.GetKey(rightMove)) - Convert.ToInt32(Input.GetKey(leftMove));
-            int ver = Convert.ToInt16(Input.GetKey(upMove)) - Convert.ToInt32(Input.GetKey(downMove));
+            int hor = Convert.ToInt16(Input.GetKey(rightMove)) - Convert.ToInt16(Input.GetKey(leftMove));
+            int ver = Convert.ToInt16(Input.GetKey(upMove)) - Convert.ToInt16(Input.GetKey(downMove));
 
             // rb.MovePosition(rb.position + new Vector2(hor, ver).normalized * speed * Time.fixedDeltaTime);
             rb.velocity = new Vector2(hor, ver).normalized * speed * (1.0f - residualVelocity.magnitude / bounceSpeed) + residualVelocity;
